+++ lib/paper_trail.rb
@@ -126,10 +126,12 @@ module PaperTrail
   # Returns PaperTrail's configuration object.
   def self.config
     @@config ||= PaperTrail::Config.instance
+    yield @@config if block_given?
+    @@config
   end
 
-  def self.configure
-    yield config
+  class << self
+    alias_method :configure, :config
   end
 end